+++ src/utils/commands/select.ts
@@ -1,0 +1,6 @@
+import * as vscode from "vscode";
+let selectCommandHandler = async () => {
+  vscode.commands.executeCommand("editor.action.smartSelect.expand");
+};
+
+export default selectCommandHandler;
