+++ src/utils/commands/show.ts
@@ -1,0 +1,5 @@
+import * as vscode from "vscode";
+let showCommandHandler = async () => {
+  vscode.commands.executeCommand("watermelon.sidebar.focus");
+};
+export default showCommandHandler;
